third_party/WebKit/LayoutTests/fast/images/style-access-during-imageChanged-crash-expected.txt

CONSOLE WARNING: Failed to decode downloaded font: data:text/plain,1
This test passes if it does not cause an assertion failure or a crash.

A 
